# llava: Add ACC OP for GPU acceleration to the Vulkan backend in the LLAVA CLIP model.